Helper update and new features
Update the helpers View and Button to be more dynamic and added a new
feature. User can now tap a button to open and close the google view
map(place holder for now) which slides in from right to left. Also added
a touch handler to handle all touch related events